+++ bulbs/super_features/metadata.py
@@ -1,7 +1,3 @@
-from collections import OrderedDict
-
-from django.utils.encoding import force_text
-
 from rest_framework import serializers
 from rest_framework.metadata import SimpleMetadata
 from rest_framework.utils.field_mapping import ClassLookupDict
@@ -10,20 +6,13 @@ from djbetty.serializers import ImageFieldSerializer
 
 from bulbs.content.serializers import AuthorField
 from bulbs.utils.fields import RichTextField
-from bulbs.infographics.data_serializers import CopySerializer, EntrySerializer, XYEntrySerializer
+from bulbs.utils.metadata import BaseSimpleMetadata
+from bulbs.infographics.data_serializers import CopySerializer, EntrySerializer
 from bulbs.super_features.serializers import BaseSuperFeatureSerializer, BaseSuperFeatureDataField
 
 
-def get_and_check_attribute(obj, attr_name):
-    attribute = getattr(obj, attr_name, None)
-    if not attribute:
-        raise AttributeError("The provided object has no '{}' attribute.".format(attr_name))
-    return attribute
+class BaseSuperFeatureMetadata(BaseSimpleMetadata):
 
-
-class BaseSuperFeatureMetadata(SimpleMetadata):
-
-    # TODO: Move fields into common file
     @property
     def label_lookup(self):
         mapping = SimpleMetadata.label_lookup.mapping
@@ -31,7 +20,6 @@ class BaseSuperFeatureMetadata(SimpleMetadata):
             AuthorField: "string",
             CopySerializer: "array",
             EntrySerializer: "array",
-            XYEntrySerializer: "array",
             ImageFieldSerializer: "image",
             RichTextField: "richtext"
         })
@@ -43,68 +31,6 @@ class BaseSuperFeatureMetadata(SimpleMetadata):
             data = self.get_custom_metadata(serializer_class, view)
             return data
         return super(BaseSuperFeatureMetadata, self).determine_metadata(request, view)
-
-    def get_label_lookup(self, field):
-        field_type = self.label_lookup[field]
-        if field_type == "field" and hasattr(field, "child_relation"):
-            return self.get_label_lookup(field.child_relation)
-        return field_type
-
-    def get_field_info(self, field):
-        """
-        This method is basically a mirror from rest_framework==3.3.3
-
-        We are currently pinned to rest_framework==3.1.1. If we upgrade,
-        this can be refactored and simplified to rely more heavily on
-        rest_framework's built in logic.
-        """
-
-        field_info = OrderedDict()
-        field_info["type"] = self.get_label_lookup(field)
-        field_info["required"] = getattr(field, "required", False)
-
-        attrs = [
-            "field_size", "read_only", "label", "help_text", "min_length", "max_length",
-            "min_value", "max_value"
-        ]
-
-        for attr in attrs:
-            value = getattr(field, attr, None)
-            if value is not None and value != "":
-                field_info[attr] = force_text(value, strings_only=True)
-
-        if getattr(field, "child", None):
-            field_info["child"] = self.get_field_info(field.child)
-        elif getattr(field, "fields", None):
-            field_info["children"] = self.get_serializer_info(field)
-
-        if (not isinstance(field, (serializers.RelatedField, serializers.ManyRelatedField)) and
-                hasattr(field, "choices")):
-            field_info["choices"] = [
-                {
-                    "value": choice_value,
-                    "display_name": force_text(choice_name, strings_only=True)
-                }
-                for choice_value, choice_name in field.choices.items()
-            ]
-
-        return field_info
-
-    def get_serializer_info(self, serializer):
-        serializer_info = super(BaseSuperFeatureMetadata, self).get_serializer_info(serializer)
-
-        if hasattr(serializer, "child"):
-            label = self.label_lookup[serializer.child]
-        else:
-            label = self.label_lookup[serializer]
-
-        # MIKE PARENT PAY ATTENTION: IDK IF THIS IS PRUDENT. LET'S DISCUSS.
-        if label != "field":
-            return OrderedDict([
-                ("type", label),
-                ("fields", serializer_info)
-            ])
-        return serializer_info
 
     def get_custom_metadata(self, serializer, view):
         fields_metadata = dict()
